Add valid data flag
I have found the temperature sensor to be a little unreliable.

I propose always returning a response and adding a `data_valid` item to the returned message.

This only covers the case where data is actually being sent. As there are a number of other occasions where data is not returned, where an error or a warn should be raised.  A second output with a status code could be used.

Cheers